Restructure input and outputs.
The input changes will make future multi-Kelly work easier to
comprehend.

The restructured output is more idiomatic for Elixir.  I initially
avoided because `:error` seemed like an overstatement of the problem.  I
am still not quite convinced, but this seems like it will make for
better (more consistent) interop with whatever sorts of third party
libraries it might encounter in the wild.

We're still not claiming a stable API, so this ought to be manageable.