Switch to multi-Kelly.
I believe this change will stablize the API.  We may add new calculators
and such, but I think the current ones will maintain their form for a
while.

This algorithm is obviously wrong!  I intend to switch this to use the
stochastic gradient ascent from Whitrow 2007.. but for now using this
very conservative approach.

The presumption at each step is that we have lost the "previous
simultaneous" stake, so our bankroll drops quickly.  This will cause us
to underbet, but most people would prefer the slight loss of return as
opposed to a huge increase in risk.

This also makes sure that we can never _actually_ overbet our entire
bankroll, as _may_ happen with, say, an EV-based approach.