Simplify the calling semantics.
Everything was getting to be a bit much and repeated, so we'll have a
standard set of options which we can extract where needed.  This will be
important for the simulator soon.

The introduction of the shared options lets us advance to 0.1.0 since
this feels like a comfortable way to continue