Naive kelly fractions.
This is simply the kelly fractions as if there were no other outstanding
bets at the same time.  Following this leads to over-betting.  In fact,
it might even sum to more than total bankroll in some cases.
Nevertheless, this is better than not having any guidance in the module.

The stotchastic gradient ascent should be up next which will scale these
appropriately.